Make packages/ui the canonical Button home; add success + accent variants
Move the Button primitive to packages/ui (shared by the shell and the embedded
plan/review apps) and add two token-driven variants for the plan app's actions:
- success: solid green (Approve), via --success
- accent:  soft accent tint (Feedback)

The frontend's components/ui/button.tsx becomes a thin re-export, so its
"@/components/ui/button" importers are unchanged and there is a single Button
implementation. Verified: ui + frontend typecheck, frontend single-file build.
